Ensure a consistent, working options data structure
Before this change, we were not so sure about the consistency of
the BugzillaQuery $options member: there were several duplicate
tests checking for 'include_fields' existence, type and correctness.

This adds a prepare_options() method which validates configuration
JSON string:
- if it's not a valid JSON, it sets an error as before,
- if it's an empty string or JSON, it sets a working default,
  so first time users get a working experience at once.

This also introduces tests for this change, to document
prepare_options() behaviour.